reorganize the vc core libraries into core, ui, and tracer
