+++ apps/VC3D/CMakeLists.txt
@@ -1,5 +1,3 @@
-# VC Main app
-set(target VC3D)
 set(CMAKE_AUTORCC ON)
 
 set(CHUNK_CACHE_SIZE_GB "10" CACHE STRING "size of the chunk cache in gigabytes")
@@ -9,7 +7,6 @@ set(srcs
     VCAppMain.cpp
     CWindow.cpp
     CWindowContextMenu.cpp
-    UDataManipulateUtils.cpp
     CVolumeViewer.cpp
     CVolumeViewerView.cpp
     SettingsDialog.cpp
@@ -46,68 +43,19 @@ if(VC_ENABLE_TSAN OR VC_ENABLE_LSAN OR VC_ENABLE_UBSAN OR VC_ENABLE_ASAN)
 endif()
 
 
-set(MACOSX_BUNDLE_ICON logo.icns)
-set_source_files_properties(${MACOSX_BUNDLE_ICON} PROPERTIES
-    MACOSX_PACKAGE_LOCATION Resources
-)
-if(APPLE)
-    list(APPEND srcs ${MACOSX_BUNDLE_ICON})
-endif()
+add_executable(VC3D ${srcs} resources.qrc)
 
-set(copyright "Copyright 2025 EduceLab")
-string(CONCAT info_string
-    ${target}
-    " Version ${PROJECT_VERSION}"
-    ", "
-    ${copyright}
-)
-
-add_executable(${target} ${srcs} resources.qrc)
-set_target_properties(${target} PROPERTIES
-    WIN32_EXECUTABLE ON
-    MACOSX_BUNDLE ON
-    MACOSX_BUNDLE_BUNDLE_NAME ${target}
-    MACOSX_BUNDLE_GUI_IDENTIFIER "com.educelab.VC"
-    MACOSX_BUNDLE_INFO_STRING ${info_string}
-    MACOSX_BUNDLE_BUNDLE_VERSION ${PROJECT_VERSION}
-    MACOSX_BUNDLE_LONG_VERSION_STRING "${target}-${CMAKE_BUILD_TYPE} Version ${PROJECT_VERSION}"
-    MACOSX_BUNDLE_SHORT_VERSION_STRING ${PROJECT_VERSION}
-    MACOSX_BUNDLE_ICON_FILE ${MACOSX_BUNDLE_ICON}
-    MACOSX_BUNDLE_COPYRIGHT ${copyright}
-)
-if(APPLE)
-    set_target_properties(${target} PROPERTIES
-        INSTALL_RPATH "@executable_path/../Frameworks/"
-    )
-endif()
-target_include_directories(${target}
-    PUBLIC
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}>
-)
-find_package(SDL2 REQUIRED)
-include_directories(${SDL2_INCLUDE_DIRS})
-find_package(GSL REQUIRED)
-include_directories(${GSL_INCLUDE_DIRS})
-target_link_libraries(${target}
-    VC::core
-    VC::metrics
-    VC::slicing
-    VC::surface
+target_link_libraries(VC3D
+    vc_core
+    vc_ui
     Qt6::Core
     Qt6::Gui
     Qt6::Widgets
-    ${OSXSecurity}
-    OpenMP::OpenMP_CXX
-    ${SDL2_LIBRARIES}
-    ${GSL_LIBRARIES}
-    z5
-    OpenMP::OpenMP_CXX
 )
 
 if (${VC_INSTALL_APPS})
     install(
-        TARGETS ${target}
+        TARGETS VC3D
         BUNDLE DESTINATION . COMPONENT Programs
         RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR} COMPONENT Programs
     )
@@ -120,7 +68,7 @@ if (${VC_INSTALL_APPS})
 
 
     qt_generate_deploy_app_script(
-        TARGET ${target}
+        TARGET VC3D
         OUTPUT_SCRIPT vc_gui_deploy_script
         NO_UNSUPPORTED_PLATFORM_ERROR
     )
