+++ core/include/vc/core/util/Surface.hpp
@@ -54,14 +54,14 @@ class PlaneSurface : public Surface
 public:
     //Surface API FIXME
     cv::Vec3f pointer() override;
-    void move(cv::Vec3f &ptr, const cv::Vec3f &offset);
+    void move(cv::Vec3f &ptr, const cv::Vec3f &offset) override;
     bool valid(const cv::Vec3f &ptr, const cv::Vec3f &offset = {0,0,0}) override { return true; };
     cv::Vec3f loc(const cv::Vec3f &ptr, const cv::Vec3f &offset = {0,0,0}) override;
     cv::Vec3f coord(const cv::Vec3f &ptr, const cv::Vec3f &offset = {0,0,0}) override;
     cv::Vec3f normal(const cv::Vec3f &ptr, const cv::Vec3f &offset = {0,0,0}) override;
     float pointTo(cv::Vec3f &ptr, const cv::Vec3f &coord, float th, int max_iters = 1000) override { abort(); };
 
-    PlaneSurface() {};
+    PlaneSurface() = default;
     PlaneSurface(cv::Vec3f origin_, cv::Vec3f normal_);
 
     void gen(cv::Mat_<cv::Vec3f> *coords, cv::Mat_<cv::Vec3f> *normals, cv::Size size, const cv::Vec3f &ptr, float scale, const cv::Vec3f &offset) override;
@@ -84,13 +84,13 @@ protected:
 class QuadSurface : public Surface
 {
 public:
-    cv::Vec3f pointer();
-    QuadSurface() {};
+    cv::Vec3f pointer() override;
+    QuadSurface() = default;
     // points will be cloned in constructor
     QuadSurface(const cv::Mat_<cv::Vec3f> &points, const cv::Vec2f &scale);
     // points will not be cloned in constructor, but pointer stored
     QuadSurface(cv::Mat_<cv::Vec3f> *points, const cv::Vec2f &scale);
-    ~QuadSurface();
+    ~QuadSurface() override;
     void move(cv::Vec3f &ptr, const cv::Vec3f &offset) override;
     bool valid(const cv::Vec3f &ptr, const cv::Vec3f &offset = {0,0,0}) override;
     cv::Vec3f loc(const cv::Vec3f &ptr, const cv::Vec3f &offset = {0,0,0}) override;
@@ -100,7 +100,7 @@ public:
     void gen(cv::Mat_<cv::Vec3f> *coords, cv::Mat_<cv::Vec3f> *normals, cv::Size size, const cv::Vec3f &ptr, float scale, const cv::Vec3f &offset) override;
     float pointTo(cv::Vec3f &ptr, const cv::Vec3f &tgt, float th, int max_iters = 1000) override;
     cv::Size size();
-    cv::Vec2f scale() const;
+    [[nodiscard]] cv::Vec2f scale() const;
 
     void save(const std::string &path, const std::string &uuid);
     void save(std::filesystem::path &path);
