+++ core/include/vc/ui/VCCollection.hpp
